+++ spec/components/breadcrumbs_spec.rb
@@ -38,6 +38,20 @@ describe "Breadcrumbs", type: :view do
     expect(structured_data["itemListElement"][2]["item"]["name"]).to eq("Section 3")
   end
 
+  it "omits breadcrumb structured data when no url is present" do
+    breadcrumbs = [
+      { title: 'Section 1', url: '/section-1' },
+      { title: 'Section 2' },
+    ]
+    structured_data = GovukPublishingComponents::Presenters::Breadcrumbs.new(breadcrumbs, "/section-3").structured_data
+    expect(structured_data["@type"]).to eq("BreadcrumbList")
+    expect(structured_data["itemListElement"].count).to eq(1)
+    expect(structured_data["itemListElement"].first["@type"]).to eq("ListItem")
+    expect(structured_data["itemListElement"].first["position"]).to eq(1)
+    expect(structured_data["itemListElement"].first["item"]["@id"]).to eq("http://www.dev.gov.uk/section-1")
+    expect(structured_data["itemListElement"].first["item"]["name"]).to eq("Section 1")
+  end
+
   it "renders all data attributes for tracking" do
     render_component(breadcrumbs: [{ title: 'Section', url: '/section' }])