+++ CHANGELOG.md
@@ -7,6 +7,10 @@
   useful summary for people upgrading their application, not a replication
   of the commit log.
 
+## Unreleased
+
+* Correct breadcrumb structured data for missing URL (PR #810)
+
 ## 16.9.1
 
 * Fix margin spacing for small form subscription links (PR #808)